Image/Server: Use single label commands struct
