+++ internal/cmd/server/server.go
@@ -38,8 +38,8 @@ func NewCommand(cli *state.State) *cobra.Command {
 		newEnableProtectionCommand(cli),
 		newDisableProtectionCommand(cli),
 		newSSHCommand(cli),
-		addLabelCmd.CobraCommand(cli.Context, hcapi2.NewClient(cli.Client()), cli),
-		removeLabelCmd.CobraCommand(cli.Context, hcapi2.NewClient(cli.Client()), cli),
+		labelCmds.AddCobraCommand(cli.Context, hcapi2.NewClient(cli.Client()), cli),
+		labelCmds.RemoveCobraCommand(cli.Context, hcapi2.NewClient(cli.Client()), cli),
 		newSetRDNSCommand(cli),
 		newAttachToNetworkCommand(cli),
 		newDetachFromNetworkCommand(cli),
